requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
django-cors-headers
importlib-metadata==5.1.0
mccabe==0.7.0
mysqlclient==2.1.1
pycodestyle==2.9.1
pyflakes==2.5.0
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.2.0
zipp==3.7.0
